+++ src/main/java/team/startup/expo/domain/application/service/FieldApplicationForTraineeService.java
@@ -1,0 +1,7 @@
+package team.startup.expo.domain.application.service;
+
+import team.startup.expo.domain.application.presentation.dto.request.ApplicationForTraineeRequestDto;
+
+public interface FieldApplicationForTraineeService {
+    void execute(String expoId, ApplicationForTraineeRequestDto dto);
+}
