+++ src/main/java/team/startup/expo/domain/application/service/PreApplicationForParticipantService.java
@@ -1,0 +1,7 @@
+package team.startup.expo.domain.application.service;
+
+import team.startup.expo.domain.application.presentation.dto.request.ApplicationForParticipantRequestDto;
+
+public interface PreApplicationForParticipantService {
+    void execute(String expoId, ApplicationForParticipantRequestDto dto);
+}
