+++ src/main/java/team/startup/expo/domain/application/service/impl/PreApplicationForParticipantServiceImpl.java
@@ -1,13 +1,13 @@
-package team.startup.expo.domain.form.service.impl;
+package team.startup.expo.domain.application.service.impl;
 
 import lombok.RequiredArgsConstructor;
 import team.startup.expo.domain.admin.Authority;
 import team.startup.expo.domain.expo.Expo;
 import team.startup.expo.domain.expo.exception.NotFoundExpoException;
 import team.startup.expo.domain.expo.repository.ExpoRepository;
-import team.startup.expo.domain.form.exception.AlreadyApplicationUserException;
-import team.startup.expo.domain.form.presentation.dto.request.ApplicationForParticipantRequestDto;
-import team.startup.expo.domain.form.service.FieldApplicationForParticipantService;
+import team.startup.expo.domain.application.exception.AlreadyApplicationUserException;
+import team.startup.expo.domain.application.presentation.dto.request.ApplicationForParticipantRequestDto;
+import team.startup.expo.domain.application.service.PreApplicationForParticipantService;
 import team.startup.expo.domain.participant.ExpoParticipant;
 import team.startup.expo.domain.participant.repository.ParticipantRepository;
 import team.startup.expo.domain.trainee.ParticipationType;
@@ -16,7 +16,7 @@ import team.startup.expo.global.annotation.TransactionService;
 
 @TransactionService
 @RequiredArgsConstructor
-public class FieldApplicationForParticipantServiceImpl implements FieldApplicationForParticipantService {
+public class PreApplicationForParticipantServiceImpl implements PreApplicationForParticipantService {
 
     private final ExpoRepository expoRepository;
     private final ParticipantRepository participantRepository;
@@ -39,7 +39,7 @@ public class FieldApplicationForParticipantServiceImpl implements FieldApplicati
                 .phoneNumber(dto.getPhoneNumber())
                 .authority(Authority.ROLE_STANDARD)
                 .attendanceStatus(false)
-                .participationType(ParticipationType.FIELD)
+                .participationType(ParticipationType.PRE)
                 .informationStatus(dto.getInformationStatus())
                 .position(dto.getPosition())
                 .expo(expo)
