+++ src/main/java/team/startup/expo/domain/application/service/impl/PreApplicationForTraineeServiceImpl.java
@@ -1,50 +1,54 @@
-package team.startup.expo.domain.form.service.impl;
+package team.startup.expo.domain.application.service.impl;
 
 import lombok.RequiredArgsConstructor;
 import team.startup.expo.domain.admin.Authority;
 import team.startup.expo.domain.expo.Expo;
 import team.startup.expo.domain.expo.exception.NotFoundExpoException;
 import team.startup.expo.domain.expo.repository.ExpoRepository;
-import team.startup.expo.domain.form.exception.AlreadyApplicationUserException;
-import team.startup.expo.domain.form.presentation.dto.request.ApplicationForParticipantRequestDto;
-import team.startup.expo.domain.form.service.PreApplicationForParticipantService;
-import team.startup.expo.domain.participant.ExpoParticipant;
+import team.startup.expo.domain.application.exception.AlreadyApplicationUserException;
+import team.startup.expo.domain.application.presentation.dto.request.ApplicationForTraineeRequestDto;
+import team.startup.expo.domain.application.service.PreApplicationForTraineeService;
 import team.startup.expo.domain.participant.repository.ParticipantRepository;
 import team.startup.expo.domain.trainee.ParticipationType;
+import team.startup.expo.domain.trainee.Trainee;
 import team.startup.expo.domain.trainee.repository.TraineeRepository;
 import team.startup.expo.global.annotation.TransactionService;
 
 @TransactionService
 @RequiredArgsConstructor
-public class PreApplicationForParticipantServiceImpl implements PreApplicationForParticipantService {
+public class PreApplicationForTraineeServiceImpl implements PreApplicationForTraineeService {
 
+    private final TraineeRepository traineeRepository;
     private final ExpoRepository expoRepository;
     private final ParticipantRepository participantRepository;
-    private final TraineeRepository traineeRepository;
 
-    public void execute(String expoId, ApplicationForParticipantRequestDto dto) {
+
+    public void execute(String expoId, ApplicationForTraineeRequestDto dto) {
         Expo expo = expoRepository.findById(expoId)
                 .orElseThrow(NotFoundExpoException::new);
 
         if (participantRepository.existsByPhoneNumberAndExpo(dto.getPhoneNumber(), expo) || traineeRepository.existsByPhoneNumberAndExpo(dto.getPhoneNumber(), expo))
             throw new AlreadyApplicationUserException();
 
-        saveParticipant(expo, dto);
+        saveTrainee(dto, expo);
     }
 
-    private void saveParticipant(Expo expo, ApplicationForParticipantRequestDto dto) {
-        ExpoParticipant expoParticipant = ExpoParticipant.builder()
-                .name(dto.getName())
-                .affiliation(dto.getAffiliation())
+    private void saveTrainee(ApplicationForTraineeRequestDto dto, Expo expo) {
+        Trainee trainee = Trainee.builder()
+                .trainingId(dto.getTrainingId())
+                .laptopStatus(dto.getLaptopStatus())
                 .phoneNumber(dto.getPhoneNumber())
-                .authority(Authority.ROLE_STANDARD)
-                .attendanceStatus(false)
+                .authority(Authority.ROLE_TRAINEE)
+                .position(dto.getPosition())
+                .schoolLevel(dto.getSchoolLevel())
+                .organization(dto.getOrganization())
+                .name(dto.getName())
                 .participationType(ParticipationType.PRE)
                 .informationStatus(dto.getInformationStatus())
-                .position(dto.getPosition())
+                .attendanceStatus(false)
                 .expo(expo)
                 .build();
 
-        participantRepository.save(expoParticipant);
+        traineeRepository.save(trainee);
     }
 }
