--- src/main/java/team/startup/expo/domain/form/service/FieldApplicationForParticipantService.java
@@ -1,7 +1,0 @@
-package team.startup.expo.domain.form.service;
-
-import team.startup.expo.domain.form.presentation.dto.request.ApplicationForParticipantRequestDto;
-
-public interface FieldApplicationForParticipantService {
-    void execute(String expoId, ApplicationForParticipantRequestDto dto);
-}
