--- src/main/java/team/startup/expo/domain/form/service/PreApplicationForTraineeService.java
@@ -1,7 +1,0 @@
-package team.startup.expo.domain.form.service;
-
-import team.startup.expo.domain.form.presentation.dto.request.ApplicationForTraineeRequestDto;
-
-public interface PreApplicationForTraineeService {
-    void execute(String expoId, ApplicationForTraineeRequestDto dto);
-}
