+++ src/main/java/team/startup/expo/domain/training/service/impl/ApplicationTrainingProServiceImpl.java
@@ -1,7 +1,7 @@
 package team.startup.expo.domain.training.service.impl;
 
 import lombok.RequiredArgsConstructor;
-import team.startup.expo.domain.form.exception.AlreadyApplicationUserException;
+import team.startup.expo.domain.application.exception.AlreadyApplicationUserException;
 import team.startup.expo.domain.sms.exception.NotFoundTraineeException;
 import team.startup.expo.domain.trainee.Trainee;
 import team.startup.expo.domain.trainee.repository.TraineeRepository;
